Lidarr Process change to artist folder name
Previously the process would only import downloads if the Lidarr Artist name, matched the downloaded artist folder name. This alters that to tell Lidarr to scan for downloads based on the actual downloaded folder name. This fixes some artists that would not previously import due to differences in the Lidarr & Folder name